Remove wallet output manipulation from transaction building lib
