+++ wallet/src/grinwallet/mod.rs
@@ -1,0 +1,29 @@
+// Copyright 2018 The Grin Developers
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Library specific to the Grin wallet implementation, as distinct from
+//! libwallet, which should build transactions without any knowledge of the
+//! wallet implementation.
+
+// TODO: Once this is working, extract a set of traits that wallet
+// implementations would need to provide
+
+#![deny(non_upper_case_globals)]
+#![deny(non_camel_case_types)]
+#![deny(non_snake_case)]
+#![deny(unused_mut)]
+#![warn(missing_docs)]
+
+pub mod selection;
+pub mod keys;
