+++ wallet/tests/transaction.rs
@@ -35,6 +35,7 @@ use chain::types::*;
 use core::{global, pow};
 use core::global::ChainTypes;
 use wallet::libwallet::{aggsig, transaction};
+use wallet::grinwallet::{keys, selection};
 use wallet::types::{OutputData, OutputStatus, WalletData};
 use util::LOGGER;
 
@@ -76,25 +77,36 @@ fn build_transaction() {
 
 	// ensure outputs we're selecting are up to date
 	let res = common::refresh_output_state_local(&wallet1.0, &wallet1.1, &chain);
+	let amount = 300_000_000_000;
+
+	// Select our outputs
+	let tx_data = selection::build_send_tx(
+		&wallet1.0,
+		&wallet1.1,
+		amount,
+		chain_tip.height,
+		3,
+		chain_tip.height,
+		1000,
+		true,
+	).unwrap();
 
 	if let Err(e) = res {
 		panic!("Unable to refresh sender wallet outputs: {}", e);
 	}
 
 	let partial_tx = transaction::sender_initiation(
-		&wallet1.0,
 		&wallet1.1,
 		&tx_id,
 		&mut sender_context_manager,
-		300_000_000_000,
 		chain_tip.height,
-		3,
-		1000,
-		true,
+		tx_data,
 	).unwrap();
 
 	let sender_context = sender_context_manager.get_context(&tx_id);
 
+	// TODO: Might make more sense to do this before the transaction
+	// building call
 	// Closure to acquire wallet lock and lock the coins being spent
 	// so we avoid accidental double spend attempt.
 	let update_sender_wallet = || {
@@ -114,7 +126,7 @@ fn build_transaction() {
 
 	// Create a potential output for this transaction
 	let (key_id, derivation) = WalletData::with_wallet(&wallet2.0.data_file_dir, |wallet_data| {
-		transaction::next_available_key(&wallet_data, &wallet2.1)
+		keys::next_available_key(&wallet_data, &wallet2.1)
 	}).unwrap();
 
 	let partial_tx = transaction::recipient_initiation(
@@ -123,7 +135,7 @@ fn build_transaction() {
 		&partial_tx,
 		&key_id,
 	).unwrap();
-	let context = recipient_context_manager.get_context(&partial_tx.id);
+	let mut context = recipient_context_manager.get_context(&partial_tx.id);
 
 	// Add the output to recipient's wallet
 	let _ = WalletData::with_wallet(&wallet2.0.data_file_dir, |wallet_data| {
@@ -131,7 +143,7 @@ fn build_transaction() {
 			root_key_id: wallet2.1.root_key_id(),
 			key_id: key_id.clone(),
 			n_child: derivation,
-			value: context.fee,
+			value: partial_tx.amount - context.fee,
 			status: OutputStatus::Unconfirmed,
 			height: 0,
 			lock_height: 0,
@@ -140,6 +152,8 @@ fn build_transaction() {
 			merkle_proof: None,
 		});
 	}).unwrap();
+	context.add_output(&key_id);
+	recipient_context_manager.save_context(context);
 
 	debug!(LOGGER, "PartialTx after step 2: recipient initiation");
 	debug!(LOGGER, "--------------------------------------------");
@@ -160,12 +174,38 @@ fn build_transaction() {
 	debug!(LOGGER, "{:?}", partial_tx);
 
 	// RECIPIENT Part 4: Recipient confirmation
+	// Get output we created in earlier step
+	let context = recipient_context_manager.get_context(&partial_tx.id);
+	let output_vec = context.get_outputs();
+	let root_key_id = &wallet2.1.root_key_id();
+
+	// operate within a lock on wallet data
+	let (key_id, derivation) = WalletData::with_wallet(&wallet2.0.data_file_dir, |wallet_data| {
+		let (key_id, derivation) = keys::retrieve_existing_key(&wallet_data, output_vec[0].clone());
+
+		wallet_data.add_output(OutputData {
+			root_key_id: root_key_id.clone(),
+			key_id: key_id.clone(),
+			n_child: derivation,
+			value: partial_tx.amount - context.fee,
+			status: OutputStatus::Unconfirmed,
+			height: 0,
+			lock_height: 0,
+			is_coinbase: false,
+			block: None,
+			merkle_proof: None,
+		});
+
+		(key_id, derivation)
+	}).unwrap();
+
 	let final_tx_recipient = transaction::finalize_transaction(
-		&wallet2.0,
 		&wallet2.1,
 		&mut recipient_context_manager,
 		&partial_tx,
 		&partial_tx,
+		&key_id,
+		derivation,
 	);
 
 	if let Err(e) = final_tx_recipient {
@@ -183,7 +223,11 @@ fn build_transaction() {
 	common::award_blocks_to_wallet(&chain, &wallet1, 3);
 
 	// Refresh wallets
-	let _ = common::refresh_output_state_local(&wallet2.0, &wallet2.1, &chain);
+	let res = common::refresh_output_state_local(&wallet2.0, &wallet2.1, &chain);
+	if let Err(e) = res {
+		panic!("Error refreshing output state for wallet: {:?}", e);
+	}
+
 	let chain_tip = chain.head().unwrap();
 	let balances = common::get_wallet_balances(&wallet2.0, &wallet2.1, chain_tip.height).unwrap();
